requirements.txt
# fastapi[all]
# pydantic
# pydantic-settings
# psycopg2-binary
# sqlalchemy
# passlib[bcrypt]
# pyjwt
# jose
# bcrypt

fastapi[all]
uvicorn
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]>=3.3.0
python-dotenv
